+++ Cnct/Cnct.Core/Configuration/CnctConfig.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Cnct.Core.Validation;
 using Newtonsoft.Json;
 
 namespace Cnct.Core.Configuration
@@ -20,12 +21,16 @@ namespace Cnct.Core.Configuration
         [JsonProperty(ItemConverterType = typeof(CnctActionConverter))]
         public ICnctActionSpec[] Actions { get; set; }
 
-        public void Validate()
+        public ConfigValidationResult Validate()
         {
-            foreach (var action in this.Actions)
+            var issues = new List<ValidationIssue>();
+
+            foreach (var action in (this.Actions ?? Array.Empty<ICnctActionSpec>()).Where(a => a != null))
             {
-                action.Validate();
+                issues.AddRange(action.Validate(this.ConfigRootDirectory));
             }
+
+            return new ConfigValidationResult(issues);
         }
 
         public async Task<bool> ExecuteAsync()